==> python/LLP Accounts 2020/0.0 LLPAccounts_UpdateDocRefsForPropertyAppExport.ipynb <==
### Fix doc_refs in db for PropertyApp export

This notebook fixes missing doc_ref columns in property_transaction table.

It iterates property_transaction records with missing doc_refs, matching up the amount, property code and date with a record from the full bank data for the year.

Requires: 
1. access to Property App VM database, 
2. spreadsheet with 6045 and 3072 data

import numpy as np
import pandas as pd

path = 'M:\My Documents\Business\Bugisiw Ltd\TaxReturn\TaxReturn_2020-2021\LLP accounts'

bank_input_file = path + '/transactions_gen.xlsx'
df6=pd.read_excel(bank_input_file,sheet_name='6045')
df6=df6[~df6.Account.isnull()]
df6=df6.astype({"Memo": str})
df3=pd.read_excel(bank_input_file,sheet_name='3072')
df3=df3[~df3.Account.isnull()]
df3=df3.astype({"Memo": str})
df6['Found']=0
df3['Found']=0

propertyidmap = {
2:'321LON',
1:'F1321LON',
14:'F2321LON',
3:'F3321LON',
4:'F4321LON',
7:'169FAW',
5:'F1169FAW',
6:'F2169FAW',
8:'F3169FAW',
73:'F1171FAW',
74:'F2171FAW',
75:'F3171FAW',
12:'163FRA',
9:'F2163FRA',
10:'F3163FRA',
11:'F4163FRA',
17:'SHOP196KIN',
15:'196AKIN',
44:'196AKIN',
16:'196BKIN',
19:'23BHAM',
34:'23CHAM',
18:'23CHAM',
31:'F58ALH',
30:'F68ALH',
33:'F78ALH',
32:'F88ALH',
26:'F1746ALH',
22:'F1846ALH',
21:'F1946ALH',
25:'F2046ALH',
28:'F2146ALH',
23:'F2246ALH',
27:'F2346ALH',
46:'F31214ALH',
52:'F41214ALH',
47:'F51214ALH',
48:'F71214ALH',
49:'F111214ALH',
53:'F101214ALH',
54:'F141214ALH',
55:'F161214ALH',
66:'F11618ALH',
67:'F31618ALH',
63:'F61618ALH',
68:'F71618ALH',
64:'F111618ALH',
71:'F121618ALH',
70:'F131618ALH',
65:'F161618ALH'
}

import sqlalchemy as sqla
import pandas as pd
import cx_Oracle
cx_Oracle.init_oracle_client(lib_dir=r"E:\dtuklaptop\e\Users\Mat\instantclient_19_9")

# Test to see if it will print the version of sqlalchemy
print(sqla.__version__)    # this returns 1.2.15 for me

# Test to see if the cx_Oracle is recognized
print(cx_Oracle.version)   # this returns 8.0.1 for me

# This fails for me at this point but will succeed after the solution described below
cx_Oracle.clientversion()

connection = cx_Oracle.connect(
    user="property_test",
    password="property_test",
    dsn="192.168.2.134/ORCL")

print("Successfully connected to Oracle Database")

Iterate property_transaction records, finding corresponding bank transactions to update doc_ref column.

cursor_query = connection.cursor()
cursor_update = connection.cursor()

query=\
'select pty.property_id, pty.alias alias, ptt.trans_id trans_id, ptt.description description, ptt.trans_date trans_date, ptt.amount amount '\
' from property_transaction ptt' \
' ,    property pty' \
' where ptt.property_id = pty.property_id' \
' and ptt.doc_ref is null'\
' and ptt.subcategory_id in (72)'\
' and ptt.trans_date >= TO_DATE(\'26/06/2019\',\'DD/MM/YYYY\')'\
' order by ptt.trans_date asc'

update = 'update property_transaction '\
         'set doc_ref = :doc_ref '\
         'where trans_id = :trans_id'

def reconcile_transactions_with_bank(df_bank, label):
    multiple_found=[]
    not_found=[]
    # For each property_transaction record, find matching record in bank download spreadsheet and update doc_ref in property_transaction
    for property_id, alias, trans_id, desc, trans_date, amount in cursor_query.execute(query):
        #print(trans_date)
        print(trans_id)
        #print(desc)
        #print(str(trans_date) + ' ' + str(trans_id) + ': ' +desc)
        p_id=propertyidmap[property_id]
        rec=df_bank.loc[(df_bank['Property']==p_id)&(df_bank['Amount']==-amount)&(df_bank['Date']==trans_date)]
        if not rec.empty:
            if(rec.shape[0]==1):
                #print(rec.iloc[0].Memo)
                doc_ref=str(label) +': ' + rec.iloc[0].Memo.replace('nan','')
                cursor_update.execute(update, [doc_ref,trans_id])
                df_bank.at[rec.iloc[0].name,'Found']=1
            else:
                cursor_update.execute(update, ['MULTIPLE',trans_id])
                multiple_found.append(trans_id)
        else:
            not_found.append(trans_id)

    #print(multiple_found)
    #print(not_found)
    return multiple_found, not_found 

#connection.commit()
#connection.rollback()

' and ptt.subcategory_id not in (89,72)'\

reconcile_transactions_with_bank(df6,'6045')

reconcile_transactions_with_bank(df3,'3072')

connection.rollback()

bank_output_file = path + '/transactions_gen_output.xlsx'
writer = pd.ExcelWriter(bank_output_file, engine='xlsxwriter',datetime_format='dd/mm/yyyy')
df6.to_excel(writer,sheet_name='6045', index=False)
df3.to_excel(writer,sheet_name='3072', index=False)
writer.save()

connection.commit()
